updated name space and project name
