+++ .github/workflows/build.yml
@@ -1,7 +1,7 @@
 # This workflow will build a .NET project
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-net
 
-name: WiFi.ts
+name: WiFits
 
 on: 
   push:
@@ -16,7 +16,7 @@ jobs:
   build:
     runs-on: macOS-latest
     env: 
-      SOLUTION_FILE: WiFi.ts.sln
+      SOLUTION_FILE: WiFits.sln
     steps:
       - uses: actions/checkout@v3
       - name: Set default Xamarin SDK versions
@@ -32,8 +32,8 @@ jobs:
 #      - name: Setup MSBuild
 #        uses: microsoft/setup-msbuild@v1.1
 
-      - name: Install dependencies for WiFi.ts
+      - name: Install dependencies for WiFits
         run: nuget restore ${SOLUTION_FILE}
 
-      - name: Build WiFi.ts
+      - name: Build WiFits
         run: msbuild ${SOLUTION_FILE} /p:Configuration=Release /t:Build /p:CreatePackage=false
