+++ README.md
@@ -1,6 +1,6 @@
-[![WiFi.ts](https://github.com/FreedomFaighter/wifi.ts/actions/workflows/build.yml/badge.svg)](https://github.com/FreedomFaighter/wifi.ts/actions/workflows/build.yml)
+[![WiFits](https://github.com/FreedomFaighter/wifi.ts/actions/workflows/build.yml/badge.svg)](https://github.com/FreedomFaighter/wifi.ts/actions/workflows/build.yml)
 
-# wifi.ts
+# wifits
 
 Originally written to record a time series of nearby wifi broadcasts for MacOS and record them in a database for diagnostic reasons typically occuring in congested wifi areas spatially
