--- WiFi.ts.sln
@@ -1,24 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WiFi.ts", "WiFi.ts\WiFi.ts.csproj", "{52CD185A-E77A-4A76-A6A0-8267E1DE92DC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{52CD185A-E77A-4A76-A6A0-8267E1DE92DC}.Debug|x86.ActiveCfg = Debug|x86
-		{52CD185A-E77A-4A76-A6A0-8267E1DE92DC}.Debug|x86.Build.0 = Debug|x86
-		{52CD185A-E77A-4A76-A6A0-8267E1DE92DC}.Release|x86.ActiveCfg = Release|x86
-		{52CD185A-E77A-4A76-A6A0-8267E1DE92DC}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		Policies = $0
-		$0.DotNetNamingPolicy = $1
-		$1.DirectoryNamespaceAssociation = PrefixedHierarchical
-		$0.StandardHeader = $2
-		$0.VersionControlPolicy = $3
-	EndGlobalSection
-EndGlobal
