--- WiFi.ts/Interfaces/IWLanModel.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Concurrent;
-
-namespace WiFi.ts
-{
-    //interface to setup a concurrent model of the network recording in the case that more then one thread is accessing the information during recording
-    public interface IWLanModel
-    {
-        ConcurrentQueue<WiFiModel> Networks { get; }
-    }
-}
