+++ WiFits/Assets.xcassets/AppIcon.appiconset/AppIcon-256.png
