+++ WiFits/Assets.xcassets/AppIcon.appiconset/AppIcon-32@2x.png
