+++ WiFits/Assets.xcassets/AppIcon.appiconset/AppIcon-512@2x.png
