+++ WiFits/Classes/WiFiModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace WiFi.ts
+namespace WiFits
 {
     public class WiFiModel : IWiFiModel
     {
