+++ WiFits/Interfaces/IWiFiDatabase.cs
@@ -1,0 +1,15 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace WiFits
+{
+
+    //interface that prepares and enqueues a database to record the wifi broadcast information
+    public interface IWiFiDatabase
+    {
+        Task CloseDB();
+        Task Enqueue(WiFiModel wiFiModel);
+        Task<bool> PrepareDB();
+        IWLanModel wLanModel { get; }
+    }
+}
