+++ WiFits/Resources/WiFits.entitlements
@@ -7,6 +7,6 @@
 	<key>com.apple.security.network.client</key>
 	<true/>
 	<key>com.apple.security.personal-information.location</key>
-	<true/>
+	<false/>
 </dict>
 </plist>
